Being A Good Open Source User
Why do open-source developers love some of their users and hate others? Learn how to submit code, patches and bug reports correctly and get open-source projects to add the features and fix the bugs that matter to you.

This talk comes out of the experience of helping run Homebrew and contributing to many other open-source projects of varying sizes.