Git: A Beginners Guide
An overview of the essentials required to interact with a Git repository. I'll explain the Git design and how it differs from other version control systems, how to submit patches to open-source projects using Git, why it is a better tool than other centralised version control systems and recommend a basic setup of a project using Github so you can share your work with others.

I'll assume a basic understanding of version control as a concept and the importance of using it for developer projects. I'll also be working mostly using command-line tools rather the GUIs to try and better explain the underlying concepts.